+++ backend/epic_app/admin.py
@@ -1,12 +1,10 @@
 from django.contrib import admin
+from epic_app.models import Answer, EpicUser, Question, Area, Group, Program
 
-from epic_app.models import Answer, EpicUser, Question
-
-# Register your models here.
+# Models exposed to the admin page .
 admin.site.register(EpicUser)
-
-# In theory this field will not be accessed as it will be a direct csv import.
+admin.site.register(Area)
+admin.site.register(Group)
+admin.site.register(Program)
 admin.site.register(Question)
-
-# These are only added until we have a more streamlined way of working.
 admin.site.register(Answer)