+++ backend/epic_app/urls.py
@@ -1,11 +1,13 @@
 from django.urls import include, path
 from rest_framework import routers
 from rest_framework.authtoken.views import obtain_auth_token
-from rest_framework.urlpatterns import format_suffix_patterns
 from epic_app import views
 
 router = routers.DefaultRouter()
 router.register(r'epicuser', views.EpicUserViewSet)
+router.register(r'area', views.AreaViewSet)
+router.register(r'group', views.GroupViewSet)
+router.register(r'program', views.ProgramViewSet)
 router.register(r'question', views.QuestionViewSet)
 router.register(r'answer', views.AnswerViewSet)
 
@@ -16,4 +18,3 @@ urlpatterns = [
     path('api/api-auth/', include('rest_framework.urls', namespace='rest_framework')),
     path('api/token-auth/', obtain_auth_token, name='api_token_auth'),
 ]
-# urlpatterns = format_suffix_patterns(urlpatterns)