+++ backend/epic_app/views.py
@@ -1,8 +1,8 @@
 # Create your views here.
 from rest_framework import viewsets, permissions
 
-from epic_app.serializers import EpicUserSerializer, QuestionSerializer, AnswerSerializer
-from epic_app.models import EpicUser, Question, Answer
+from epic_app.serializers import EpicUserSerializer, QuestionSerializer, AnswerSerializer, AreaSerializer, GroupSerializer, ProgramSerializer
+from epic_app.models import EpicUser, Question, Answer, Area, Group, Program
 
 class EpicUserViewSet(viewsets.ModelViewSet):
     """
@@ -13,6 +13,39 @@ class EpicUserViewSet(viewsets.ModelViewSet):
     """
     queryset = EpicUser.objects.all().order_by('username')
     serializer_class = EpicUserSerializer
+    permission_classes = [permissions.DjangoModelPermissions]
+
+class AreaViewSet(viewsets.ReadOnlyModelViewSet):
+    """
+    Default view set for 'Area'
+
+    Args:
+        viewsets (ModelViewSet): Derives directly from ModelViewSet
+    """
+    queryset = Area.objects.all().order_by('name')
+    serializer_class = AreaSerializer
+    permission_classes = [permissions.DjangoModelPermissions]
+
+class GroupViewSet(viewsets.ReadOnlyModelViewSet):
+    """
+    Default view set for 'Group'
+
+    Args:
+        viewsets (ModelViewSet): Derives directly from ModelViewSet
+    """
+    queryset = Group.objects.all().order_by('name')
+    serializer_class = GroupSerializer
+    permission_classes = [permissions.DjangoModelPermissions]
+
+class ProgramViewSet(viewsets.ReadOnlyModelViewSet):
+    """
+    Default view set for 'Program'
+
+    Args:
+        viewsets (ModelViewSet): Derives directly from ModelViewSet
+    """
+    queryset = Program.objects.all().order_by('name')
+    serializer_class = ProgramSerializer
     permission_classes = [permissions.DjangoModelPermissions]
 
 class QuestionViewSet(viewsets.ReadOnlyModelViewSet):